Desktop/robat/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
aiosqlite==0.19.0


